requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
openai===1.40.6
pydantic==2.8.2